Move export format to a separate file
